Deleted .DS_Store
What even is this file? From what I see is that it is a Mac OS file for file/folder config, either way, if it is not need in the project then it should be added to the .gitignore file we have or should make so that if anyone on Mac OS uses VS Code or something to push the code via git, this wouldn't happen.